fix(services/hf): move config enums out of config
